# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MamboFades Barber API
 * OpenAPI spec version: 0.1.0
 */
export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface LoginInput {
  password: string;
}

export interface AuthStatus {
  authenticated: boolean;
}

export interface HealthStatus {
  status: string;
}

export interface Style {
  id: number;
  name: string;
  description: string;
  price: number;
  durationMinutes: number;
  /** @nullable */
  imageUrl: string | null;
  /** e.g. fades, cuts, designs, beard */
  category: string;
  isActive: boolean;
  createdAt?: string;
}

export interface StyleInput {
  name: string;
  description: string;
  price: number;
  durationMinutes: number;
  /** @nullable */
  imageUrl?: string | null;
  category: string;
  isActive?: boolean;
}

export interface StyleUpdate {
  name?: string;
  description?: string;
  price?: number;
  durationMinutes?: number;
  /** @nullable */
  imageUrl?: string | null;
  category?: string;
  isActive?: boolean;
}

export interface Booking {
  id: number;
  clientName: string;
  clientPhone: string;
  clientEmail: string;
  styleId: number;
  styleName: string;
  /** YYYY-MM-DD */
  bookingDate: string;
  /** HH:MM (24hr) */
  bookingTime: string;
  /** pending | confirmed | cancelled | completed */
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface BookingInput {
  clientName: string;
  clientPhone: string;
  clientEmail: string;
  styleId: number;
  bookingDate: string;
  bookingTime: string;
  /** @nullable */
  notes?: string | null;
}

export interface BookingUpdate {
  status?: string;
  bookingDate?: string;
  bookingTime?: string;
  /** @nullable */
  notes?: string | null;
}

export interface Summary {
  totalBookings: number;
  pendingBookings: number;
  confirmedBookings: number;
  todayBookings: number;
  totalStyles: number;
  /** @nullable */
  popularStyle: string | null;
}

export type ListBookingsParams = {
/**
 * Filter by date (YYYY-MM-DD)
 * @nullable
 */
date?: string | null;
/**
 * Filter by status (pending, confirmed, cancelled, completed)
 * @nullable
 */
status?: string | null;
};

export type GetAvailableSlotsParams = {
/**
 * Date in YYYY-MM-DD format
 */
date: string;
};