# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MamboFades Barber API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: styles
    description: Haircut style catalogue
  - name: bookings
    description: Appointment bookings
  - name: summary
    description: Business summary
  - name: auth
    description: Admin authentication
  - name: Storage
    description: Object storage upload and serving endpoints.
paths:
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.

  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"
        "401":
          description: Invalid password

  /auth/logout:
    post:
      operationId: adminLogout
      tags: [auth]
      summary: Admin logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"

  /auth/me:
    get:
      operationId: getAuthMe
      tags: [auth]
      summary: Check current auth status
      responses:
        "200":
          description: Auth status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"

  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /styles:
    get:
      operationId: listStyles
      tags: [styles]
      summary: List all haircut styles in catalogue
      responses:
        "200":
          description: List of styles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Style"
    post:
      operationId: createStyle
      tags: [styles]
      summary: Add a new style to the catalogue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StyleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Style"

  /styles/{id}:
    get:
      operationId: getStyle
      tags: [styles]
      summary: Get a single style
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Style found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Style"
        "404":
          description: Not found
    patch:
      operationId: updateStyle
      tags: [styles]
      summary: Update a style
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StyleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Style"
    delete:
      operationId: deleteStyle
      tags: [styles]
      summary: Delete a style
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all bookings
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: ["string", "null"]
          description: Filter by date (YYYY-MM-DD)
        - name: status
          in: query
          required: false
          schema:
            type: ["string", "null"]
          description: Filter by status (pending, confirmed, cancelled, completed)
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Validation error

  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a single booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    delete:
      operationId: deleteBooking
      tags: [bookings]
      summary: Cancel/delete a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /bookings/available-slots:
    get:
      operationId: getAvailableSlots
      tags: [bookings]
      summary: Get available time slots for a given date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
          description: Date in YYYY-MM-DD format
      responses:
        "200":
          description: Available slots
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /summary:
    get:
      operationId: getSummary
      tags: [summary]
      summary: Get business summary stats
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Summary"

components:
  schemas:
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    LoginInput:
      type: object
      required: [password]
      properties:
        password:
          type: string

    AuthStatus:
      type: object
      required: [authenticated]
      properties:
        authenticated:
          type: boolean

    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Style:
      type: object
      required: [id, name, description, price, durationMinutes, imageUrl, category, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        durationMinutes:
          type: integer
        imageUrl:
          type: ["string", "null"]
        category:
          type: string
          description: e.g. fades, cuts, designs, beard
        isActive:
          type: boolean
        createdAt:
          type: string

    StyleInput:
      type: object
      required: [name, description, price, durationMinutes, category]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        durationMinutes:
          type: integer
        imageUrl:
          type: ["string", "null"]
        category:
          type: string
        isActive:
          type: boolean

    StyleUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        durationMinutes:
          type: integer
        imageUrl:
          type: ["string", "null"]
        category:
          type: string
        isActive:
          type: boolean

    Booking:
      type: object
      required: [id, clientName, clientPhone, clientEmail, styleId, styleName, bookingDate, bookingTime, status, createdAt]
      properties:
        id:
          type: integer
        clientName:
          type: string
        clientPhone:
          type: string
        clientEmail:
          type: string
        styleId:
          type: integer
        styleName:
          type: string
        bookingDate:
          type: string
          description: YYYY-MM-DD
        bookingTime:
          type: string
          description: HH:MM (24hr)
        status:
          type: string
          description: pending | confirmed | cancelled | completed
        notes:
          type: ["string", "null"]
        createdAt:
          type: string

    BookingInput:
      type: object
      required: [clientName, clientPhone, clientEmail, styleId, bookingDate, bookingTime]
      properties:
        clientName:
          type: string
        clientPhone:
          type: string
        clientEmail:
          type: string
        styleId:
          type: integer
        bookingDate:
          type: string
        bookingTime:
          type: string
        notes:
          type: ["string", "null"]

    BookingUpdate:
      type: object
      properties:
        status:
          type: string
        bookingDate:
          type: string
        bookingTime:
          type: string
        notes:
          type: ["string", "null"]

    Summary:
      type: object
      required: [totalBookings, pendingBookings, confirmedBookings, todayBookings, totalStyles, popularStyle]
      properties:
        totalBookings:
          type: integer
        pendingBookings:
          type: integer
        confirmedBookings:
          type: integer
        todayBookings:
          type: integer
        totalStyles:
          type: integer
        popularStyle:
          type: ["string", "null"]