# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MamboFades Barber API
 * OpenAPI spec version: 0.1.0
 */

export interface Booking {
  id: number;
  clientName: string;
  clientPhone: string;
  clientEmail: string;
  styleId: number;
  styleName: string;
  /** YYYY-MM-DD */
  bookingDate: string;
  /** HH:MM (24hr) */
  bookingTime: string;
  /** pending | confirmed | cancelled | completed */
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}