# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MamboFades Barber API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  clientName: string;
  clientPhone: string;
  clientEmail: string;
  styleId: number;
  bookingDate: string;
  bookingTime: string;
  /** @nullable */
  notes?: string | null;
}