# lib/api-zod/src/generated/types/style.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MamboFades Barber API
 * OpenAPI spec version: 0.1.0
 */

export interface Style {
  id: number;
  name: string;
  description: string;
  price: number;
  durationMinutes: number;
  /** @nullable */
  imageUrl: string | null;
  /** e.g. fades, cuts, designs, beard */
  category: string;
  isActive: boolean;
  createdAt?: string;
}