# lib/api-zod/src/generated/types/styleInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MamboFades Barber API
 * OpenAPI spec version: 0.1.0
 */

export interface StyleInput {
  name: string;
  description: string;
  price: number;
  durationMinutes: number;
  /** @nullable */
  imageUrl?: string | null;
  category: string;
  isActive?: boolean;
}